# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HMCA API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pms/reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Get all reservations",
                "operationId": "c4505ccccb423747e2ad7e6463c2bd75",
                "responses": {
                    "200": {
                        "description": "List of reservations fetched successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Create a new reservation",
                "operationId": "7466a2488ae1451e27ec0af547e521c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "room_id",
                                    "check_in",
                                    "check_out",
                                    "first_name"
                                ],
                                "properties": {
                                    "room_id": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-11-07 12:00:00"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-11-08 12:00:00"
                                    },
                                    "arrival_from": {
                                        "type": "string",
                                        "example": "Chennai"
                                    },
                                    "booking_type": {
                                        "type": "string",
                                        "example": "walk-in"
                                    },
                                    "booking_reference_no": {
                                        "type": "string",
                                        "example": "BK12345"
                                    },
                                    "purpose_of_visit": {
                                        "type": "string",
                                        "example": "Business"
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Early check-in requested"
                                    },
                                    "adults": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "children": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "+91"
                                    },
                                    "mobile_no": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Mr"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "father_name": {
                                        "type": "string",
                                        "example": "Robert Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "occupation": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "dob": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-06-12"
                                    },
                                    "anniversary": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2020-02-14"
                                    },
                                    "nationality": {
                                        "type": "string",
                                        "example": "Indian"
                                    },
                                    "is_vip": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "contact_type": {
                                        "type": "string",
                                        "example": "personal"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "India"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Tamil Nadu"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Chennai"
                                    },
                                    "zipcode": {
                                        "type": "string",
                                        "example": "600001"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Mount Road, Chennai"
                                    },
                                    "identity_type": {
                                        "type": "string",
                                        "example": "aadhaar"
                                    },
                                    "identity_no": {
                                        "type": "string",
                                        "example": "1234-5678-9123"
                                    },
                                    "identity_comments": {
                                        "type": "string",
                                        "example": "ID verified at front desk"
                                    },
                                    "front_doc": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "back_doc": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "guest_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "discount_reason": {
                                        "type": "string",
                                        "example": "Corporate discount"
                                    },
                                    "discount_percent": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "10"
                                    },
                                    "commission_percent": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "5"
                                    },
                                    "commission_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "500"
                                    },
                                    "payment_mode": {
                                        "type": "string",
                                        "example": "card"
                                    },
                                    "advance_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "2000"
                                    },
                                    "advance_remarks": {
                                        "type": "string",
                                        "example": "Paid via credit card"
                                    },
                                    "booking_charge": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "4000"
                                    },
                                    "tax": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "480"
                                    },
                                    "service_charge": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "200"
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "4680"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "reserved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
...
        "/api/pms/reservations/{id}/check-in": {
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Mark reservation as checked-in",
                "operationId": "cceb678711ab48e000b70eed0468b258",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation checked in successfully"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            }
        },
        "/api/pms/reservations/{id}/check-out": {
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Mark reservation as checked-out (room set to dirty)",
                "operationId": "cd516358d1b4a5760ea45071631956cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation checked out successfully"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            }
        },
        "/api/pms/reservations/{id}/cancel": {
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Cancel a reservation",
                "operationId": "a4337c6bffb4d42c0359b60270088fce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reservation UUID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation cancelled successfully"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}